# backend/apps/api/src/main/kotlin/dk/example/feedback/controller/FeedbackController.kt
package dk.example.feedback.controller

import dk.example.feedback.dto.FeedbackEventDto
import dk.example.feedback.dto.SubmitFeedbackResponseDto
import dk.example.feedback.payloads.StartFeedbackEventInput
import dk.example.feedback.payloads.SubmitFeedbackInput
import dk.example.feedback.service.FeedbackService
import io.swagger.v3.oas.annotations.tags.Tag
import org.springframework.security.access.prepost.PreAuthorize
import org.springframework.security.core.annotation.AuthenticationPrincipal
import org.springframework.security.oauth2.jwt.Jwt
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RestController

@RestController
@Tag(name = "Feedback")
@RequestMapping("/feedback")
class FeedbackController(val feedbackService: FeedbackService) {

    @PreAuthorize("isAuthenticated()")
    @PostMapping("/start")
    fun startFeedbackEvent(
        @RequestBody startFeedbackEventInput: StartFeedbackEventInput,
        @AuthenticationPrincipal principal: Jwt
    ): FeedbackEventDto {
        return feedbackService.startEvent(pinCode = startFeedbackEventInput.pinCode, jwt = principal)
    }

    @PreAuthorize("isAuthenticated()")
    @PostMapping("/submit")
    fun submitFeedback(
        @RequestBody input: SubmitFeedbackInput,
        @AuthenticationPrincipal principal: Jwt
    ): SubmitFeedbackResponseDto {
        return feedbackService.submitFeedback(
            feedbackInputList = input.feedback,
            pinCode = input.pinCode,
            jwt = principal
        )
    }
}