# contracts/openapi/feedback-api.yaml
openapi: 3.1.0
info:
  title: Feedback API
  description: API documentation for Lets Grow application.
  version: v1
servers:
- url: http://localhost:18080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /account:
    put:
      tags:
      - Account
      operationId: modifyAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyAccountInput"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
      - Account
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountInput"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - Account
      operationId: deleteAccount
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /account/role:
    put:
      tags:
      - Account
      operationId: updateRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleInput"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /account/fcm-token:
    put:
      tags:
      - Account
      operationId: linkFCMTokenToAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkFCMTokenToAccountInput"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /feedback/submit:
    post:
      tags:
      - Feedback
      operationId: submitFeedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackInput"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitFeedbackResponseDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /feedback/start:
    post:
      tags:
      - Feedback
      operationId: startFeedbackEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartFeedbackEventInput"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackEventDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /account/logout:
    post:
      tags:
      - Account
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutInput"
        required: true
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /bootstrap:
    get:
      tags:
      - Bootstrap
      operationId: getBootstrap
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bootstrap/bootstrap-update/{hash}:
    get:
      tags:
      - Bootstrap
      summary: Get bootstrap update by hash
      description: Returns a full bootstrap payload when the provided hash is stale;
        returns 204 when unchanged.
      operationId: getBoostrapUpdate
      parameters:
      - name: hash
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Bootstrap updated; full bootstrap payload returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapDto"
        "204":
          description: Bootstrap hash unchanged; no response body.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ActivityDto:
      type: object
      description: "Manager-facing activity with configuration, active questions,\
        \ event history, and trend analytics."
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier for the activity.
        title:
          type: string
          description: Display title shown to managers and participants.
        agenda:
          type: string
        owner:
          $ref: "#/components/schemas/OwnerDto"
        runMode:
          type: string
          description: Run mode that controls how events are joined and processed.
          enum:
          - MANUAL
          - AUTOMATIC
        sendEmails:
          type: boolean
        invitedEmails:
          type: array
          items:
            type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventDto"
        currentQuestions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionDto"
        trend:
          $ref: "#/components/schemas/ActivityTrendDto"
      required:
      - currentQuestions
      - events
      - id
      - invitedEmails
      - owner
      - runMode
      - sendEmails
      - title
      - trend
    ActivityTrendDto:
      type: object
      description: Activity trend computed from comparable event ratings normalized
        to the 0-5 scale.
      properties:
        direction:
          type: string
          description: Trend direction across comparable events.
          enum:
          - improving
          - stable
          - declining
          - insufficient_data
        indicator:
          type: string
          description: UI indicator derived from the trend direction.
          enum:
          - positive
          - neutral
          - negative
        metric:
          type: string
          description: Metric used to compute the trend.
          enum:
          - average_rating
        latestValue:
          type: number
          format: double
        previousValue:
          type: number
          format: double
        delta:
          type: number
          format: double
        comparedEventCount:
          type: integer
          format: int32
          description: Number of events used for trend comparison.
      required:
      - comparedEventCount
      - direction
      - indicator
      - metric
    EventDto:
      type: object
      description: "Manager-facing event summary with schedule, join details, and\
        \ question snapshot."
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier for the event.
        date:
          type: string
          format: date-time
          description: Scheduled start timestamp for the event.
        durationInMinutes:
          type: integer
          format: int32
        location:
          type: string
        pinCode:
          type: string
          description: Participant join code for submitting feedback.
        createdFromMailListener:
          type: boolean
        calendarProvider:
          type: string
          enum:
          - GOOGLE
          - APPLE
          - MICROSOFT
          - ZOOM
        calendarEventId:
          type: string
        averageRating:
          type: number
          format: double
          description: Average rating normalized to 0-5 when comparable feedback exists.
        overallFeedbackSummary:
          $ref: "#/components/schemas/OverallFeedbackSummaryDto"
        questionsSnapshot:
          type: array
          items:
            $ref: "#/components/schemas/QuestionDto"
      required:
      - createdFromMailListener
      - date
      - durationInMinutes
      - id
      - questionsSnapshot
...
    OverallFeedbackSegmentationStatsDto:
      type: object
      description: Percentage distribution of emoji feedback outcomes in a event summary.
      properties:
        verySadPercentage:
          type: number
          format: double
        sadPercentage:
          type: number
          format: double
        happyPercentage:
          type: number
          format: double
        veryHappyPercentage:
          type: number
          format: double
      required:
      - happyPercentage
      - sadPercentage
      - veryHappyPercentage
      - verySadPercentage
    OverallFeedbackSummaryDto:
      type: object
      description: Aggregated feedback summary for a event across all questions.
      properties:
        segmentationStats:
          $ref: "#/components/schemas/OverallFeedbackSegmentationStatsDto"
        countStats:
          $ref: "#/components/schemas/OverallFeedbackCountStatsDto"
        unseenResponses:
          type: integer
          format: int32
        responses:
          type: integer
          format: int32
      required:
      - countStats
      - responses
      - segmentationStats
      - unseenResponses
    OwnerDto:
      type: object
      description: Owner identity attached to activities and events.
      properties:
        id:
          type: string
          description: Account identifier for the owner.
        name:
          type: string
        email:
          type: string
      required:
      - id
    QuestionDto:
      type: object
      description: Question reference used in activity and event snapshots.
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier for the question.
        text:
          type: string
          description: Question text shown to participants.
      required:
      - id
      - text
...
    ModifyAccountInput:
      type: object
      description: Input payload for updating account profile information.
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
    UpdateRoleInput:
      type: object
      description: Input payload for changing the authenticated account role.
      properties:
        role:
          type: string
          description: Target role for the account.
      required:
      - role
    LinkFCMTokenToAccountInput:
      type: object
      description: Input payload to link an FCM token to the authenticated account.
      properties:
        fcmToken:
          type: string
          description: Firebase Cloud Messaging token for push notifications.
      required:
      - fcmToken
    FeedbackInput:
      type: object
      description: Single feedback answer submitted for one question in a event.
      properties:
        comment:
          type: string
        emoji:
          type: string
          enum:
          - VerySad
          - Sad
          - Happy
          - VeryHappy
        thumbsUpThumpsDown:
          type: string
          enum:
          - Up
          - Down
        opinion:
          type: string
          enum:
          - StronglyDisagree
          - Disagree
          - Neutral
          - Agree
          - StronglyAgree
        zeroToTen:
          type: integer
          format: int32
        questionId:
          type: string
          format: uuid
          description: Question identifier this feedback answer targets.
        feedbackType:
          type: string
          description: Feedback type that determines which answer field is expected.
          enum:
          - Emoji
          - Comment
          - ThumpsUpThumpsDown
          - Opinion
          - ZeroToTen
      required:
      - feedbackType
      - questionId
    SubmitFeedbackInput:
      type: object
      description: Input payload for submitting participant feedback to a event.
      properties:
        feedback:
          type: array
          description: Collection of feedback answers keyed by question id.
          items:
            $ref: "#/components/schemas/FeedbackInput"
        pinCode:
          type: string
          description: Event pin code the feedback belongs to.
      required:
      - feedback
      - pinCode
    OwnerInfoDto:
      type: object
      description: Owner profile metadata shown to participants.
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
    ParticipantEventDto:
      type: object
      description: Participant-visible event payload including join and feedback status.
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier for the event.
        date:
          type: string
          format: date-time
          description: Scheduled start timestamp for the event.
        pinCode:
          type: string
          description: Pin code used to join the event.
        durationInMinutes:
          type: integer
          format: int32
        location:
          type: string
        createdFromMailListener:
          type: boolean
        ownerInfo:
          $ref: "#/components/schemas/OwnerInfoDto"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantQuestionDto"
        feedbackSubmited:
          type: boolean
        recentlyJoined:
          type: boolean
      required:
      - createdFromMailListener
      - date
      - durationInMinutes
      - feedbackSubmited
      - id
      - ownerInfo
      - questions
      - recentlyJoined
    ParticipantQuestionDto:
      type: object
      description: Participant-facing question payload for feedback submission.
      properties:
        id:
          type: string
          format: uuid
          description: Question identifier used when submitting feedback.
        questionText:
          type: string
          description: Question text presented to the participant.
        feedbackType:
          type: string
          description: Feedback format expected for this question.
          enum:
          - Emoji
          - Comment
          - ThumpsUpThumpsDown
          - Opinion
          - ZeroToTen
      required:
      - feedbackType
      - id
      - questionText
    SubmitFeedbackResponseDto:
      type: object
      description: Response returned after participant feedback submission.
      properties:
        shouldPresentRatingPrompt:
          type: boolean
          description: Signals whether the client should show an app-rating prompt.
        event:
          $ref: "#/components/schemas/ParticipantEventDto"
      required:
      - event
      - shouldPresentRatingPrompt
    StartFeedbackEventInput:
      type: object
      description: Input payload for joining a feedback event by pin code.
      properties:
        pinCode:
          type: string
          description: Event pin code entered by the participant.
      required:
      - pinCode
    FeedbackEventDto:
      type: object
      description: Participant-facing event payload returned after joining by pin
        code.
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantQuestionDto"
        ownerInfo:
          $ref: "#/components/schemas/OwnerInfoDto"
        date:
          type: string
          format: date-time
          description: Scheduled timestamp for the joined event.
      required:
      - date
      - ownerInfo
      - questions
...
    CreateAccountInput:
      type: object
      description: Input payload for account creation during first authentication.
      properties:
        requestedRole:
          type: string
          description: Requested role for the new account.
        fcmToken:
          type: string
    AccountInfoDto:
      type: object
      description: Authenticated account profile metadata.
      properties:
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
    BootstrapDto:
      type: object
      description: Bootstrap response payload for authenticated clients.
      properties:
        role:
          type: string
          description: Resolved role for the authenticated account.
        accountInfo:
          $ref: "#/components/schemas/AccountInfoDto"
        participantEvents:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantEventDto"
        managerData:
          $ref: "#/components/schemas/ManagerDataDto"
      required:
      - accountInfo
      - participantEvents
    EmojiQuestionFeedbackSummary:
      type: object
      description: Distribution summary for emoji feedback.
      properties:
        countVerySad:
          type: integer
          format: int32
        countSad:
          type: integer
          format: int32
        countHappy:
          type: integer
          format: int32
        countVeryHappy:
          type: integer
          format: int32
        percentageVerySad:
          type: number
          format: double
        percentageSad:
          type: number
          format: double
        percentageHappy:
          type: number
          format: double
        percentageVeryHappy:
          type: number
          format: double
      required:
      - countHappy
      - countSad
      - countVeryHappy
      - countVerySad
      - percentageHappy
      - percentageSad
      - percentageVeryHappy
      - percentageVerySad
    ManagerDataDto:
      type: object
      description: Manager-only bootstrap section with activities and analytics.
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityDto"
        notificationHistory:
          $ref: "#/components/schemas/NotificationHistoryDto"
        bootstrapHash:
          type: string
          format: uuid
          description: Hash used by clients to detect manager data updates.
        questionAnalytics:
          type: array
          items:
            $ref: "#/components/schemas/ManagerQuestionAnalyticsDto"
      required:
      - activities
      - bootstrapHash
      - notificationHistory
      - questionAnalytics
    ManagerQuestionAnalyticsDto:
      type: object
      description: Per-question analytics aggregated across manager events.
      properties:
        questionId:
          type: string
          format: uuid
          description: Canonical question identifier across event snapshots.
        questionText:
          type: string
          description: Latest question text used for the canonical question.
        feedbackType:
          type: string
          description: Feedback format for this question.
          enum:
          - Emoji
          - Comment
          - ThumpsUpThumpsDown
          - Opinion
          - ZeroToTen
        eventCount:
          type: integer
          format: int32
        responseCount:
          type: integer
          format: int32
        latestAskedAt:
          type: string
          format: date-time
          description: Timestamp when this question was most recently asked.
        overallSummary:
          $ref: "#/components/schemas/QuestionFeedbackSummaryDto"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/QuestionTrendPointDto"
      required:
      - eventCount
      - feedbackType
      - questionId
      - questionText
      - responseCount
      - timeline
    NotificationHistoryDto:
      type: object
      description: Manager notification history summary for newly received feedback.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NotificationHistoryItem"
        unseenTotal:
          type: integer
          format: int32
          description: Total unseen notification items for the manager.
      required:
      - items
      - unseenTotal
    NotificationHistoryItem:
      type: object
      description: Single notification history item linked to an activity event.
      properties:
        id:
          type: string
          format: uuid
          description: Stable identifier for the notification history item.
        date:
          type: string
          format: date-time
          description: Timestamp when the notification was created.
        eventTitle:
          type: string
        eventId:
          type: string
          format: uuid
          description: Identifier of the related event or event.
        newFeedbackCount:
          type: integer
          format: int32
        seenByManager:
          type: boolean
      required:
      - date
      - eventId
      - eventTitle
      - id
      - newFeedbackCount
      - seenByManager
    OpinionQuestionFeedbackSummary:
      type: object
      description: Distribution summary for agree/disagree opinion feedback.
      properties:
        countStronglyAgree:
          type: integer
          format: int32
        countAgree:
          type: integer
          format: int32
        countStronglyDisagree:
          type: integer
          format: int32
        countDisagree:
          type: integer
          format: int32
        percentageStronglyAgree:
          type: number
          format: double
        percentageAgree:
          type: number
          format: double
        percentageStronglyDisagree:
          type: number
          format: double
        percentageDisagree:
          type: number
          format: double
      required:
      - countAgree
      - countDisagree
      - countStronglyAgree
      - countStronglyDisagree
      - percentageAgree
      - percentageDisagree
      - percentageStronglyAgree
      - percentageStronglyDisagree
    QuestionFeedbackSummaryDto:
      type: object
      description: Question feedback summary where only one subtype is populated per
        feedback type.
      properties:
        emojiQuestionFeedbackSummary:
          $ref: "#/components/schemas/EmojiQuestionFeedbackSummary"
        thumpsQuestionFeedbackSummary:
          $ref: "#/components/schemas/ThumpsQuestionFeedbackSummary"
        opinionQuestionFeedbackSummary:
          $ref: "#/components/schemas/OpinionQuestionFeedbackSummary"
        zeroToTenQuestionFeedbackSummary:
          $ref: "#/components/schemas/ZeroToTenQuestionFeedbackSummary"
    QuestionTrendPointDto:
      type: object
      description: Single event trend point for question analytics.
      properties:
        eventId:
          type: string
          format: uuid
          description: Identifier of the event represented by this trend point.
        eventDate:
          type: string
          format: date-time
          description: Event timestamp represented by this trend point.
        responseCount:
          type: integer
          format: int32
        summary:
          $ref: "#/components/schemas/QuestionFeedbackSummaryDto"
      required:
      - eventDate
      - eventId
      - responseCount
    ThumpsQuestionFeedbackSummary:
      type: object
      description: Distribution summary for thumbs up/down feedback.
      properties:
        countUp:
          type: integer
          format: int32
        countDown:
          type: integer
          format: int32
        percentageUp:
          type: number
          format: double
        percentageDown:
          type: number
          format: double
      required:
      - countDown
      - countUp
      - percentageDown
      - percentageUp
...
    LogoutInput:
      type: object
      description: Input payload for account logout and device token unlink.
      properties:
        fcmToken:
          type: string
          description: Firebase Cloud Messaging token to remove from the account.
      required:
      - fcmToken
    ApiError:
      properties:
        timestamp:
          type: string
        message:
          type: string
        domainCode:
          type: string
          enum:
          - FEEDBACK_ALREADY_SUBMITTED
          - EVENT_ALREADY_JOINED
          - CANNOT_JOIN_OWN_EVENT
          - CANNOT_GIVE_FEEDBACK_TO_SELF
          - PINCODE_NOT_FOUND
        exceptionType:
          type: string
        path:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT